Double check we are still loading the right image
This fixes a case where we’ve fully hydrated an image and we’re ready
to render it on the image view, but we have to do this asynchronously.

If the cell went offscreen and then got reused in this time (which might happen if the user is scrolling really fast) then we’ll just ignore the callback if the URL doesn’t match.