Standardize code formatting (#497)
Modernise Swift Code
Created configs for SwiftFormat and applied it to the codebase. This
mostly tidied up a lot of whitespace and applied consistent indentation.
It also removed unused variables and redundant calls to self and
return statements. Imports and switch cases are now ordered
alphabetically.

Used the Preview macro in all cases except for one.

Applied a standard header comment from SwiftFormat, which reduces noise
ta the top of the file. The history of who created what file when can be
found in git.

Added an IDETemplateMacros.plist file to the project so that new files
will have this new standard header.

SwiftFormat(https://github.com/nicklockwood/SwiftFormat) has not been
added as a dependency. It's easier for people to just run it locally
before committing. Or somebody could periodically run it over the entire
codebase to keep things tidy.

Remove AppearanceProxies file
It only had an extension on UILabel, and there are none left in the app.
Sort by state abbreviation
Since the keys come first it makes sense to sort by those rather than full names